+++ src/main/java/models/Secretary.java
@@ -1,7 +1,7 @@
 package models;
 /** Secretary class that extends User */
 public class Secretary extends User {
-    public Secretary (int userID, String name, String password, String email, String role) {
+    public Secretary (long userID, String name, String password, String email, String role) {
         super(userID, name, password, email, role);
 
     }
